requirements.txt
Django
django-rest-framework
requests
beautifulsoup4
scrapy
psycopg2-binary
python-dotenv
